// integrations/ukg-pro/types.ts
export interface EmployeeDetails {
  firstName: string;
  lastName: string;
  preferredName: string;
  emailAddress: string;
  countryCode: string;
  languageCode: string;
  employeeNumber: number;
  employeeId: string;
  personId: string;
  userIntegrationKey: string;
  companyName: string;
  companyId: string;
  supervisorId: string;
  salaryOrHourly: "S" | "H";
  fullTimeOrPartTime: "F" | "P";
  isActive: boolean;
  workLocation: string;
  jobCode: string;
  projectCode: string;
  orgLevel1Code: string;
  orgLevel2Code: string;
  orgLevel3Code: string;
  orgLevel4Code: string;
  middleName: string;
  workPhone: string;
  homePhone: string;
  employeeAddress1: string;
  employeeAddress2: string;
  city: string;
  state: string;
  zipCode: string;
  terminationDate: string;
  hireDate: string;
  supervisorName: string;
  prefix: string;
  suffix: string;
  alternateEmailAddress: string;
  gender: "M" | "F" | "U";
  employeeStatus: string;
  employeeType: string;
  emplStatusStartDate: string;
  dateInJob: string;
  dateOfLastHire: string;
  jobGroupCode: string;
  alternateJobTitle: string;
}